Removed the "I" prefixes from all the interfaces.
All implementors will now start with Simple.
